// database.types.ts
export type Database = {
  public: {
    Tables: {
      parking_restrictions: {
        Row: {
          approved_at: string | null
          approved_by: string | null
          "Controlled Parking Zone": string | null
          id: string
          "Image URL": string | null
          "Last Updated": string | null
          "Last Uploaded": string | null
          Latitude: number
          Location: unknown | null
          Longitude: number
          "Maximum Stay": string | null
          "Nearest Machine": string | null
          Notes: string | null
          "Parking Spaces": number | null
          Postcode: string | null
          "Restriction Type": string | null
          "Road Name": string | null
          Source: string | null
          status: Database["public"]["Enums"]["restriction_status"]
          submitted_at: string | null
          submitted_by: string | null
          "Times Of Operation": string | null
          "Valid Parking Permits": string | null
        }
        Insert: {
          approved_at?: string | null
          approved_by?: string | null
          "Controlled Parking Zone"?: string | null
          id?: string
          "Image URL"?: string | null
          "Last Updated"?: string | null
          "Last Uploaded"?: string | null
          Latitude: number
          Location?: unknown | null
          Longitude: number
          "Maximum Stay"?: string | null
          "Nearest Machine"?: string | null
          Notes?: string | null
          "Parking Spaces"?: number | null
          Postcode?: string | null
          "Restriction Type"?: string | null
          "Road Name"?: string | null
          Source?: string | null
          status?: Database["public"]["Enums"]["restriction_status"]
          submitted_at?: string | null
          submitted_by?: string | null
          "Times Of Operation"?: string | null
          "Valid Parking Permits"?: string | null
        }
        Update: {
          approved_at?: string | null
          approved_by?: string | null
          "Controlled Parking Zone"?: string | null
          id?: string
          "Image URL"?: string | null
          "Last Updated"?: string | null
          "Last Uploaded"?: string | null
          Latitude?: number
          Location?: unknown | null
          Longitude?: number
          "Maximum Stay"?: string | null
          "Nearest Machine"?: string | null
          Notes?: string | null
          "Parking Spaces"?: number | null
          Postcode?: string | null
          "Restriction Type"?: string | null
          "Road Name"?: string | null
          Source?: string | null
          status?: Database["public"]["Enums"]["restriction_status"]
          submitted_at?: string | null
          submitted_by?: string | null
          "Times Of Operation"?: string | null
          "Valid Parking Permits"?: string | null
        }
        Relationships: []
      }
      spatial_ref_sys: {
        Row: {
          auth_name: string | null
          auth_srid: number | null
          proj4text: string | null
          srid: number
          srtext: string | null
        }
        Insert: {
          auth_name?: string | null
          auth_srid?: number | null
          proj4text?: string | null
          srid: number
          srtext?: string | null
        }
        Update: {
          auth_name?: string | null
          auth_srid?: number | null
          proj4text?: string | null
          srid?: number
          srtext?: string | null
        }
        Relationships: []
      }
      users: {
        Row: {
          created_at: string | null
          email: string
          id: string
          name: string | null
          role: Database["public"]["Enums"]["user_role"]
        }
        Insert: {
          created_at?: string | null
          email: string
          id: string
          name?: string | null
          role?: Database["public"]["Enums"]["user_role"]
        }
        Update: {
          created_at?: string | null
          email?: string
          id?: string
          name?: string | null
          role?: Database["public"]["Enums"]["user_role"]
        }
        Relationships: []
      }
    }
// ...
    Enums: {
      restriction_status: "approved" | "pending"
      user_role: "admin" | "user"
    }
    CompositeTypes: {
      geometry_dump: {
        path: number[] | null
        geom: unknown | null
      }
      valid_detail: {
        valid: boolean | null
        reason: string | null
        location: unknown | null
      }
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never